ml/requirements.txt
transformers
numpy
timm
torch
torchvision
onnxruntime
pillow
scipy
onnxscript
onnxruntime
onnxruntime-tools